Use dev_server to determine the url of the devserver.
The dynamic_suite code assumed the devserver variable in the autotest
config was a string that represented the url for the devserver.
This is no longer a correct assumption as it can be a list of devservers.
This change modifies the dynamic_suite logic to use a DevServer instance
to tell it the URL it should use for the devserver.

BUG=chromium-os:26451
TEST=Unittests

Change-Id: I0bc63029630d9e884a23b16ccdf75da1756e6518
Reviewed-on: https://gerrit.chromium.org/gerrit/28447
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>
Commit-Ready: Chris Sosa <sosa@chromium.org>