[autotest] Fix logic around finding reimaging hosts to lock
The existing logic could, in the case where a job had only HostQueueEntries
with no hosts in them, wind up calling get_hosts([], status='Running'), which
would return ALL currently running hosts -- not just the Running hosts
associated with the currently interesting reimaging job.  This would lead
to all those hosts getting locked, which is A Bad Thing.

This change fixes this logic so that, if we don't actually have any hosts
to check on, we just skip all that and loop back around.

BUG=chromium-os:32907
TEST=unit
TEST=run_suite with all applicable hosts locked.

Change-Id: I2403c7bf0a64b011853558c0a26bc85923870286
Reviewed-on: https://gerrit.chromium.org/gerrit/28491
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>