[autotest] Add |num| argument to create_suite_job RPC
This CL does not make any suite control file honor the argument yet.
It also does not provide a way to pass this arg from run_suite.py.

This change should drop in safely, and be a no-op to the rest of the
infrastructure.  As the other changes are more likely to impact
behavior in production, I wanted to get this in on its own so we don't
churn this API a bunch with reverts and re-lands.

BUG=chromium-os:33027
TEST=use atest create suite to create a suite with the new -n arg; check the job on the AFE to see that your value got injected into the suite control file.

Change-Id: I53431b565ec6bf664baaadeca8ddd279a500f381
Reviewed-on: https://gerrit.chromium.org/gerrit/28651
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>