Fix USB related FAFT components
It turns out that FAFT does not properly connect USB mux to the host
(does not enable the mux explicitly). This causes FAFT components
requiring USB flash stick access pass if the mux happens to be enabled
and fail if it does not.

This change ensures that the mux is enabled when required.

BUG=chrome-os-partner:12022
TEST=manual
  . start servo in a separate session:
    $ sudo servod -c servo_link_overlay.xml

  . in another session disable the mux and try running a test
    requiring USB flash stick access:

    $ dut-control usb_mux_sel1:dut_sees_usbkey  usb_mux_oe1:off
    $ ./run_remote_tests.sh --board=link   --remote=<ip addr> --servo \
         CorruptBothKernelAB/control.normal
    $ ./run_remote_tests.sh --board=link   --remote=<ip addr> --servo \
        InvalidUSB

  Observe both tests succeed (they were failing before this change)

Change-Id: I666997dd16e2ad064f4c5781e36e1a612da51a0d
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28769