Fix security_SeccompSyscallFilters TRAP_handler test case.
The test has the SIGSYS handler modify a static variable to check
whether the handler is being executed. However, the compiler is most
likely messing with the order of statements since it cannot know that
the static variable is being modified in the handler.

Use 'volatile' to tell the compiler that the variable might change
from under its feet.

BUG=chromium-os:32567
TEST=security_SeccompSyscallFilters passes on lumpy.
TEST=security_SeccompSyscallFilters compiles on x86-alex and tegra2_kaen.

Change-Id: Iafeb45462d61cac969ed0db6ecd4132c58e94ae7
Reviewed-on: https://gerrit.chromium.org/gerrit/28816
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Kees Cook <keescook@chromium.org>
Commit-Ready: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Jim Hebert <jimhebert@chromium.org>