[autotest] Fix unittest for JobStatus
testWaitForResults was actually never checking to be sure that the test
statuses we inject were really being reported by job_status.wait_for_results()
This was because wait_for_results() was actually side-effecting the list of
jobs passed in and emptying it out as it ran.  Thus, we'd get to the end of the
test and loop through 'jobs'...which was now an empty list!

This isn't biting us in practice because the real code never looks at the list
of jobs again after it passes it into wait_for_results().  This CL fixes the
code to use a local copy of the job list, as well.

BUG=None
TEST=unit
TEST=use run_suite.py to run a real suite

Change-Id: Ib35188502803b6ea4828d9514e419e900d8f999b
Reviewed-on: https://gerrit.chromium.org/gerrit/28840
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>