Test spurious contacts when panel power is cycled
This patch checks whether there are any spurious contacts when the
touch screen panel power is cycled.

BUG=chrome-os-partner:12212
TEST=./run_remote_tests.sh --remote=$MACHINE_IP hardware_TouchScreenPowerCycles
The panel power will be cycled 5 times.
When the screen is turned on, you may try to tap the screen, and
observe that it reports failure.

Change-Id: Ie6588272baed1d4972b318d9b8bf8bccb4bd00f7
Reviewed-on: https://gerrit.chromium.org/gerrit/29155
Reviewed-by: Yufeng Shen <miletus@chromium.org>
Reviewed-by: Tai-Hsu Lin <sheckylin@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>