faft: Changing FORCE_DEV_SWITCH_ON flag requires reboot to get a clear state
The GBB flag FORCE_DEV_SWITCH_ON is a work-around for factory to force to
boot into a dev mode. It overwrites the actual dev switch state which is
stored in TPM. So the system may or may not in the dev mode. When we
disable this flag, the next reboot mode is unknown, either normal mode or
dev mode. So it is safe to reboot it to come up with a clear state.

BUG=chromium-os:32874
TEST=Run firmware_TryFwB/control.dev passed on a new installed firmware.

Change-Id: I32486324562e6b8847c9b88de7bab386773fa647
Reviewed-on: https://gerrit.chromium.org/gerrit/29593
Reviewed-by: Vic Yang <victoryang@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>