faft: Change the file structures of the control files for FAFT tests
The current control files for FAFT tests is:
 - control: symbolic link to control.both;
 - control.both: run the test under both normal mode and dev mode;
 - control.normal: run the test under normal mode;
 - control.dev: run the test under dev mode.

Only the control.both is marked as the suite faft. When we call
run_remote_tests.sh with suite:faft. Both control (symbolic link)
and control.both will be included. So each test will be run twice.
Some errors will be reported.

So this CL is to change the control files into:
 - control: run the test under normal mode;
 - control.dev: run the test under dev mode.

Both control and control.dev are marked as the suite faft.

BUG=chromium-os:32859
TEST=run_remote_tests.sh --board=daisy --remote=$REMOTE_IP suite:faft

Change-Id: I53bfe3f7baa37effd75d1691280c33d58c503890
Reviewed-on: https://gerrit.chromium.org/gerrit/29741
Reviewed-by: Vic Yang <victoryang@chromium.org>
Reviewed-by: Mike Truty <truty@chromium.org>
Reviewed-by: Rajesh Chenna <rchenna@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>