Modify dynamic_suite to pass along the user of the suite run.
Traditionally dynamic_suite would only mark the original suite job
as being owned by the user that kicked off the suite. This produced the
unfortunate scenario where the user running the suite now has the same
access rights to machines as the main automation user chromeos-test.

This also hides the sub jobs from the user making it harder for them
to see exactly what is going on under the covers.

With a move to start supporting try-jobs it is important to show the user
all of their jobs and also create a separation between what machines users
can run on and what machines the automated testing can run on.

TEST=utils/unittest_suite.py server.cros.dynamic_suite_unittest
BUG=chromium-os:33534
Change-Id: Ide831e4abef695cbd8b5d3d3b714d3c7c65f4c5a
Reviewed-on: https://gerrit.chromium.org/gerrit/30273
Tested-by: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Ready: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>