Autotest: Release webserver port for new monitor_db process.
Currently monitor_db_babysitter is responsible for launching monitor_db
and relaunching in case it crashes. However if the status server port is
not properly released this can cause the monitor_db to continually loop
trying to launch monitor_db and failing.

In order to work around this issue the babysitter will first check if
the status server port is open, and if not open it before launching
monitor_db.

BUG=chromium-os:32961
TEST=launched monitor_db_babysitter when the port was held and when it
was free and made sure monitor_db lunched successfully.

Change-Id: I36b9107b6158e9e1d2cac83d45c07a5325c1ecec
Reviewed-on: https://gerrit.chromium.org/gerrit/30323
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Ready: Simran Basi <sbasi@google.com>
Tested-by: Simran Basi <sbasi@google.com>