[autotest] Call old signal handler when leaving SignalsAsExceptions.
Autotest relies on a SIGTERM handler to close the pidfile and kill other
processes in the process group, so we can't just skip calling it.
However, we do need it to not run in the middle of any code within the
scope of SignalsAsExceptions.

BUG=chromium-os:33523
TEST=pylint, unit

Change-Id: I0e814180264c78c2708c95490323c749eb6252e3
Reviewed-on: https://gerrit.chromium.org/gerrit/30452
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Ready: Alex Miller <milleral@chromium.org>