Revert "[autotest] Locked hosts should not count as usable"
This reverts commit 91351a637853bd5986f19b36da92d7bb6fa5cc73

The semantics of this change would mean that any device in the lab that's happily being reimaged for another test is counted as unavailable.  While it is _right now_ it's fine, and will be usable in a bit.  So, reverting until we get the logic right.

Change-Id: I5e7ec59c3b8b76e6f7c6033ee1c1588c7d625d76
Reviewed-on: https://gerrit.chromium.org/gerrit/30760
Commit-Ready: Chris Masone <cmasone@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>