Disable cleanup pathway in gs_offloader.py
After 6 months of being in place we have hit an issue where we were
unable to diagnose a bug due to this. Now that we are on a much larger
set of machines and have a more stable offloading process the cleanup
command will be disabled.

Follow up will happen on crosbug.com/26784

Bug affected by this cleanup: crbug.com/140408

BUG=chromium-os:26784
TEST=Syntax

Change-Id: If552dc2c0cc3be541dd5507c9f1620028bdf9cde
Reviewed-on: https://gerrit.chromium.org/gerrit/30766
Reviewed-by: Simran Basi <sbasi@google.com>
Commit-Ready: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Scott Zawalski <scottz@chromium.org>