Add a script to bind the MySQL process to a CPU set
This script creates a specified CPU set and bind a running MySQL process
to the set. When running with option "-r", the script restarts MySQL
first before performing the binding. Hence this script should act as a
wrapper script for starting MySQL if we want to limit the CPU usage of
MySQL.

BUG=chromium-os:33582
TEST=local tests

Change-Id: I48b0ea9b838e0de3569c429c3b4fccede20f67ae
Reviewed-on: https://gerrit.chromium.org/gerrit/30977
Commit-Ready: Yu-Ju Hong <yjhong@chromium.org>
Reviewed-by: Yu-Ju Hong <yjhong@chromium.org>
Tested-by: Yu-Ju Hong <yjhong@chromium.org>