Set LIBC_FATAL_STDERR_=1 in the Host environment variables.
Arrange for commands executed on the host to include setting
LIBC_FATAL_STDERR_ in the process' environment variables.  The
setting will provide additional debug output for certain errors of
interest to security tests, to the extent that callers that run
commands include stderr in the logs.

BUG=chromium-os:31430
TEST=log the output of `env` using a custom client-side test

Change-Id: I5efd8d569fbe29e101285cd74dbaeb004d6c3a4e
Reviewed-on: https://gerrit.chromium.org/gerrit/30921
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Jim Hebert <jimhebert@chromium.org>
Reviewed-by: Kees Cook <keescook@chromium.org>
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>