[autotest] Refactor email sending out of suite_scheduler.
Sending warnings/errors as emails is a useful thing to be able to do
outside of suite_scheduler. This exposes the ability to create a logging
handler from a config file and a subject line that sends emails for a
certain severity level of logging messages.

BUG=chromium-os:32925
TEST=pylint, ran unittests
TEST=ran suite_scheduler.py and verified it sends emails

Change-Id: I4a678779ea06989d0f86a89017b4a6753afe257e
Reviewed-on: https://gerrit.chromium.org/gerrit/30348
Commit-Ready: Alex Miller <milleral@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Tested-by: Alex Miller <milleral@chromium.org>