faft: Implement a new test for Software Sync EC update
This test verified the EC software sync. The EC binary is embedded in
the BIOS image. On AP startup, AP verifies the EC by comparing the
hash of the current EC with the hash of the embedded EC image. So
updating EC is simple and we just need to change the EC binary on
the BIOS image.

The test steps are:
 1. check EC is RO boot and AP is RO boot,
    extract the original EC binary on current AP RW A and save it locally,
    copy the new_ec binary to the EC area on the current AP RW A,
    disable RO_BOOT preamble flag,
    reboot;
 2. check EC is RW boot and AP is RW A boot,
    compare the current EC with the new_ec binary,
    copy the original EC binary to the EC area on the current AP RW A,
    reboot;
 3. check EC is RW boot and AP is RW A boot,
    compare the current EC with the original EC binary,
    enable RO_BOOT preamble flag,
    reboot;
 4. check EC is RO boot and AP is RW A boot.

BUG=chrome-os-partner:12997
TEST=After merging some EC update test changes, run:
run_remote_test.sh -a "new_ec=ec_autest_image.bin" firmware_UpdateECBin/control$

Software sync still has some issue (http://crosbug.com/p/12680) that RO-Boot ->
TwoStop update is failed. So this is expected to be failed.

Change-Id: I0801266f5554129c17fc39312cfcb3163650f227
Reviewed-on: https://gerrit.chromium.org/gerrit/31328
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>