[autotest] Add job=id-owner entries to suite job keyvals for infra issues.
The dynamic suite infrastructure adds keyvals to the suite job to
allow us to map each test back to the job that ran it when we're
generating result reports.  We recently started adding results to the
suite when we detect SERVER_JOB and CLIENT_JOB failures.  As we don't
know about these at job-scheduling time, and add a result for them unless
there's an issue, we didn't have a keyval entry for them.

This CL adds one.

BUG=chromium-os:33956
TEST=run_suite.py, and ensure one job fails when it's trying to copy results back.  Check to make sure that the failure is reported, and that the emitted link points to the right job.

Change-Id: I2d00cb6875e80942ca31c343bfbeff284d072a2d
Reviewed-on: https://gerrit.chromium.org/gerrit/31717
Commit-Ready: Chris Masone <cmasone@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>