[autotest] Filter out debian control files.
When caching the list of control files from the file system or
devserver, strip out any control files that end in 'src/debian/control'.
This way, we can't mistake files that are a part of the debian packaging
system as files that are a part of the autotest system.

BUG=chromium-os:33916
TEST=ran local dummy suite, verified warnings about src/debian/control
     files no longer occurred.

Change-Id: I9273bc94c753bca8f43d7f27335a723b10c6e128
Reviewed-on: https://gerrit.chromium.org/gerrit/32260
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Alex Miller <milleral@chromium.org>