faft: Restore to the original BIOS when fail on firmware_UpdateECBin
We used the system shellball to recovery the BIOS when fail. It just makes
the DUT able to work later but loses the original BIOS if it is different
with the shellball one.

The safe way is to keep the original BIOS at setup() phase. When test fails,
the cleanup() phase is executed. If the SHA of BIOS RW A changes, the test
restores to the original BIOS.

BUG=chrome-os-partner:12997
TEST=Force to make firmware_UpdateECBin fail and see:
16:19:35 INFO | ======== Running FAFT sequence step 2 ========
16:19:35 INFO | calling <function <lambda> at 0x22a2758>
16:20:03 INFO | BIOS changed. Restore the original BIOS...

Change-Id: I846568c674f9fad223f34b71cb1cfb8e36539ed4
Reviewed-on: https://gerrit.chromium.org/gerrit/32371
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>