firmware_TouchpadMTB: add a validator to verify no gaps in a line
Let distance be the distance between two consecutive points.
When we see a gap in a tracking line, it generally means that
current distance is significantly larger than the previous
distance and next distance. In this CL, a new validator is added
to verify that there are no gaps in a tracking line.

BUG=chromium-os:34265
TEST=When running firmware_TouchpadMTB test, watch NoGapValidator
grade in the result frame on the right hand side of the window.

Change-Id: I0be51510936d5e751fde693d67f614ec16e20a46
Reviewed-on: https://gerrit.chromium.org/gerrit/32698
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>