Autotest: Verifying Loop in Scheduler
If the scheduler dies between setting a host queue entry to Verifying
and actually creating the special task to do the Verify, it will
prevent the scheduler from launching thus stopping it from processing
jobs/tests.

This change removes the code that explicitly raises an exception for
this situation and instead changes the job's status to Queued so that
it can be processed.

BUG=chromium-os:33990
TEST=Changed the scheduler code so that it dies after changing the
     host_queue_entry status to Verifying and ensured the scheduler
     could not launch. Then added my fix and it began working, job
     was processed and I could still assign new future jobs to this
     host.

Change-Id: Iab60579911b0a82c90bbd23f14eab546e4a04e3a
Reviewed-on: https://gerrit.chromium.org/gerrit/32610
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>