[autotest] Filter out non-running HostQueueEntries while watching jobs
When watching a reimaging job to see what hosts it gets assigned, so
we can lock them, we need to be sure that an assigned host that's in
the Running state is actually running the reimaging job we care about.
The best way to do this seems to be to pre-filter any HostQueueEntries
that are not, themselves, Running.  This means we won't even look at a
host until we know that it's doing something for the job we care
about.

BUG=chromium-os:34294
TEST=unit

Change-Id: I508168e198f2286537911dfcfa0e815f710c3732
Reviewed-on: https://gerrit.chromium.org/gerrit/32850
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>