firmware_TouchpadMTB: Handle the case when there are no variations
This CL handles the special case when there are no variations for
a gesture.

The prompt of every test is printed in two places. At the first
place, it is printed as text with escaped color code in the final
test report, which is called color_prompt. At the second place,
it is printed to the window as regular text which is called
monochrome_prompt. The window subsystem will then encode the prompt
with colors in html markup language. This CL fixed the bug by
setting monochrome_prompt when there are no variations.

BUG=chromium-os:34115
TEST=None. Will be tested when the gestures without variations are
added later.

Change-Id: Ie8b0247f6d3ccc0a5a9625b07987f145e18b70b0
Reviewed-on: https://gerrit.chromium.org/gerrit/32890
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>