firmware_TouchpadMTB: add a validator to measure reversed motions
If the finger tracking direction is horizontal, check if there are
any reversed motions on x axis. Likewise, check y axis for vertical
lines. Check both x and y axes for diagonal lines. In this CL, a
new validator is added to measure the reversed motions.

BUG=chromium-os:34266
TEST=Run "DISPLAY=:0 sudo python main.py" in crosh of a chromebook.
The first test should be one_finger_tracking. Watch the test result
of NoReversedMotionValidator on the right hand side of the window.

Change-Id: Ic9f1d20b17299e2f98888511ac2549c89cee96fe
Reviewed-on: https://gerrit.chromium.org/gerrit/32910
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>