firmware_TouchpadMTB: add a validator to verify the number of packets.
When making a swipe, there should be a minimal number of packets
reported so that the gestures library could work correctly. In this CL,
a new validator is added to verify the number of reported packets
should satisfy the criteria.

BUG=chromium-os:34267
TEST=Run "DISPLAY=:0 sudo python main.py" in crosh of a chromebook.
Watch the test result of 'one_finger_swipe' and 'two_finger_swipe'.
Watch the test result of CountPacketsValidator on the right hand side
of the window.

Change-Id: I7a47e7baa0e3b65229cba45faed363eb17e36fa1
Reviewed-on: https://gerrit.chromium.org/gerrit/32913
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>