firmware_TouchpadMTB: add a validator to verify pinch to zoom gesture
For 'pinch to zoom in' gesture, the fingers draw farther, while for
'pinch to zoom out' gesture, the fingers draw closer. In either case,
check that the distance between two fingers become larger/smaller by
at least the specified distance in the criteria string.

BUG=chromium-os:34268
TEST=Run "DISPLAY=:0 sudo python main.py" in crosh of a chromebook.
Watch the test result of 'pinch_to_zoom' gesture which is shown in
CountPacketsValidator on the right hand side of the window.

Change-Id: I836862bd33061806a18c98afc926ea4323943a81
Reviewed-on: https://gerrit.chromium.org/gerrit/33069
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>