faft: Bundle SAFT libraries with Autotest
The SAFT was installed into the Chrome OS test image on build time. But the
Autotest code are copied to DUT on runtime. It always makes unsynced when
rapid changes, i.e. the Autotest code is new but the SAFT libraries are old.

This CL is to bundle the latest SAFT libraries with Autotest. So both
Autotest and SAFT libraries are copied on runtime.

The SAFT code keeps unchanged in this CL.

The top changes of the latest SAFT:
a740928 saft: Fix the argument and the return value of get_internal_disk()
80fbf0e saft: Modify ecbin size field on the DTB blob
48459e9 Fetch internal disk when using removable disk
5420e9b saft: Write the modified firmware body to flashrom after updating ecbin
18ef555 saft: Fix bug that the offset of the ecbin should be 4-byte aligned

BUG=chromium-os:7891
TEST=run_remote_tests.sh --remote=$DUT control.faft_lv1
     run_remote_tests.sh --remote=$DUT control.faft_lv2
     run_remote_tests.sh --remote=$DUT control.faft_lv3

Change-Id: I97f1c87b2a7e1b18a4b87a450e62aadf1d78886d
Reviewed-on: https://gerrit.chromium.org/gerrit/33178
Reviewed-by: Vic Yang <victoryang@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>