Add a "coming soon" link to autotest dashboard for new perf graphs.
This change modifies the autotest dashboard UI to display a "COMING SOON"
link in the dashboard header (underneath the "last updated:" message).
The link points to a proof-of-concept UI for a revamped perf dashboard
that is meant to replace the existing chromeOS perf dashboard soon.
The new perf dashboard more closely follows the style used by the Chrome
perf graphs.

As soon as the transition is made from the current perf dashboard UI to
the new one, the COMING SOON link will be removed.

BUG=None
TEST=None

Change-Id: Iecfaf84b22dbab387610a997443b45d741e3cd34
Reviewed-on: https://gerrit.chromium.org/gerrit/33253
Reviewed-by: Nirnimesh <nirnimesh@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Dennis Jeffrey <dennisjeffrey@chromium.org>
Tested-by: Dennis Jeffrey <dennisjeffrey@chromium.org>