faft: Use normal full power-off process for the tests which change cgpt value
In normal cases, we use the hardware warm reboot after each step. It caused
an issue that some tests change a cgpt value but the new value seems to be
cached in the block device, not fast enough to write the physical disk.

This CL changes the reboot action to use the normal full power-off process,
which presses the power-button. The system shutdown process should write all
cached data to the physical disk. It takes longer time to reboot but it acts
like more realistic user behavior.

BUG=chrome-os-partner:14055
TEST=Run firmware_CorruptKernelB passed on Snow.

Change-Id: Id5c1ada4b727cc48eb80f4b6820a795db05f424f
Reviewed-on: https://gerrit.chromium.org/gerrit/33268
Reviewed-by: Vic Yang <victoryang@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>