firmware_TouchpadMTB: fix a bug about getting (x, y) in a slot
The evdev maintains a state machine and only reports the change.
Some cases to pay attention to:

  . If the pressure of the target slot is changed but (x, y) is
    not changed, it will only report the pressure.
  . If nothing is changed in the target slot, but the pressure,
    x, or y is changed in another slot, it will report nothing
    about the target slot.
  . Slot 0 may or may not be reported depending on the state
    before the finger touching the pad.

In every case above, we need to report (x, y) for the target
slot. This CL fixes mtb.MTB.get_x_y() to behave in this way.
Some unittests are added to verify the change.

BUG=chromium-os:34435
TEST=Run the unittests.
$ cd .../autotest/files/client/site_tests/firmware_TouchpadMTB
$ PYTHONPATH=.:../../bin/input python tests/mtb_unittest.py
The unit tests should all pass.

Change-Id: Iab0e2b5d3da597e24d15a4d00cd2c04b5ed2c2f2
Reviewed-on: https://gerrit.chromium.org/gerrit/33295
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>