firmware_TouchpadMTB: add a new validator to detect drumroll
The drumroll cases occur in platforms lumpy and link exhibit
the same phenomenon -- the movement of positions in the same
slot is significantly larger than normally it should be as taps.
More specifically, the drumroll may happen in the case that the
reported positions line up in a line between the two tap groups.
(When a user make quick two-finger taps, there are roughly two
tap groups.) It could also happen that the reported points in a
slot jump back and forth between the two tap groups. In either
case, to check the max movement in a slot is a simple way to
detect drumroll.

This CL extracts the points for every trackid ID, and verifies
that the max movement in the tracking ID should be within a
threshold value specified in the criterion string.

BUG=chromium-os:34270
TEST=Run the unittests.
$ cd .../autotest/files/client/site_tests/firmware_TouchpadMTB
$ PYTHONPATH=.:../../bin/input python tests/mtb_unittest.py
The unit tests should all pass.

Change-Id: I8b8c9a1d186c7b166a3ad1ebd3d83e81da7a9d92
Reviewed-on: https://gerrit.chromium.org/gerrit/33407
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>