[autotest] Use |None| placeholder for hosts not in |Running| status.
If we do not append anything in |get_all_hostnames()|, then we return an
empty list of hosts to lock. We start off with an empty list of hosts
being locked. Therefore, we're trivially done!

What we should be doing is using |None| as a placeholder for hosts that
we know about, but aren't ready to be locked yet.

This also removes the extra logging that was put in place to help
diagnose this issue.

BUG=None
TEST=local run_suite

Change-Id: Id0d2ad7965fd05aa98dc583bcbf9cc17991db950
Reviewed-on: https://gerrit.chromium.org/gerrit/33358
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Alex Miller <milleral@chromium.org>