firmware_TouchpadMTB: Speed up dump screenshots
This CL tries to reduce the screenshot dumping time by tuning
the parameters of imagemagic import command utility. With
some experiments of tuning the parameters, it is possible to
optimize the dumping speed. First, it is faster to crop the
root window rather than assigning the mtplot window id to the
utility. Secondly, by tuning the -quality parameter, we could
get a good balance between speed and quality. There is no any
noticeable quality reduction since mtplot image is a simple
black background with some color dots. Experiments on link
show that the dumping latency is cut in half with the new
parameters.

BUG=chromium-os:34565
TEST=None. Test this CL with next one to see the speed up.

Change-Id: I9f1f124fa93abc002da5154bb69cc2034db27963
Reviewed-on: https://gerrit.chromium.org/gerrit/33589
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>