firmware_TouchpadMTB: tune timeout parameter to speed up the tests
The timeout value was originally set to 500 ms after the user
performing a gesture and there is no new event data coming in. This
is unnecessarily long for most gestures, e.g., one-finger-tracking.
The typical frame inter-arrival time of a touchpad is about
1000 ms / 80 = 12.5 ms. It would be sufficient to set the default
timeout as 12.5 ms * 10 = 125 ms. This could reduce the waiting time
for showing the test result after a user performing a gesture.

For the gestures such as drumroll in which fingers need to leave the
pad temporarily, we set the timeout to the original 500 ms to make
sure that it does not timeout prematurely.

BUG=chromium-os:34565
TEST=Run "DISPLAY=:0 sudo python main.py" in crosh of a chromebook.
After performing a gesture, the test result is shown on the right
hand side of the window in a significantly faster speed.

Change-Id: I04b9fbd7424d562d8d92b0cff9c85db4ebff36b1
Reviewed-on: https://gerrit.chromium.org/gerrit/33590
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>